Move guessing game components to their own file
